add more longlongs

+++ error_kmers.py
@@ -320,8 +320,8 @@ def correct_sam_test(samfile, conf_regions, outfile, tcounts, perror, kgraph):
                     print("pe1:",pe1)
                     print("p:",p)
                     raise
-                A[j] = np.array([[1.0 - pe1, pe1],[pe0 - pe0*pe1, 1.0 - pe0+pe0*pe1]]) #A is size len(counts), but A[0] is meaningless
-                E[j] = np.array([[p_a_given_note[count]],[p_a_given_e[count]]]) #E is size len(counts)
+                A[j] = np.array([[1.0 - pe1, pe1],[pe0 - pe0*pe1, 1.0 - pe0+pe0*pe1]], dtype=np.longlong) #A is size len(counts), but A[0] is meaningless
+                E[j] = np.array([[p_a_given_note[count]],[p_a_given_e[count]]], dtype=np.longlong) #E is size len(counts)
 
             A = np.array(t_baum_welch(A, E, pi))
             for j, count in enumerate(counts):
